Reduced BMNG/Landsat num levels to 10. Repaired data size calculation that was overflowing int.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@1810 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWindIOS/TAIGA/TAIGA/BulkRetrieverController.h
@@ -20,7 +20,7 @@
     UIView* startAccessory;
     UIView* stopAccessory;
     UIProgressView* progress;
-    int dataSize;
+    double dataSize;
     WWBulkRetriever* retriever;
 }
